Adds a diagnosis prompt generation tool.
This is essentially a text wrapper but has a >99% return rate for ONLY the Diagnostic Report. This also creates a small json artifact with the issue id. The file Metadata along with the artifact can be used to track usages for monitoring resource spikes and the like. This fits in with the current plans to add python functionality and allows those nodes with python or bash>python3 functionality to use this tool immediately for stable Diagnostic Reports on any error thrown.